Move app to app.js
